requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
constantly==23.10.4
crispy-bootstrap5==2025.6
cryptography==45.0.5
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.4
django-crispy-forms==2.4
djangorestframework==3.16.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.5.1
msgpack==1.1.1
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==6.2.0
scikit-learn==1.7.1
scikit-surprise==1.1.4
scipy==1.15.3
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
watchdog==6.0.0
whitenoise==6.9.0
zope.interface==7.2
